refactor : 코드 수정
 - Controller 생성
 - render method 수정
 - Position Class method 순서 수정
 - Piece Class getFile, getRank -> getXPosition, getYPosition수정
