+++ src/main/resources/templates/index.html
@@ -293,7 +293,7 @@
             error: getPieceInfoError,
             success: getPieceInfoSuccess
         });
-        alert("새로운게임이 시작되었습니다. 새로고침하여 시작하세요.");
+        alert("새로운게임이 시작되었습니다.");
     });
 
     function getPieceInfoError() {
